+++ bauh/gems/arch/__init__.py
@@ -6,6 +6,7 @@ from bauh.commons import resource
 
 ROOT_DIR = os.path.dirname(os.path.abspath(__file__))
 BUILD_DIR = '{}/arch'.format(TEMP_DIR)
+PACKAGE_CACHE_DIR = '{}/pkg_cache'.format(BUILD_DIR)
 ARCH_CACHE_PATH = CACHE_PATH + '/arch'
 CATEGORIES_FILE_PATH = ARCH_CACHE_PATH + '/categories.txt'
 URL_CATEGORIES_FILE = 'https://raw.githubusercontent.com/vinifmor/bauh-files/master/arch/categories.txt'